Let's further improve. (See commit details below).
lol I totally realized that the ViewStats is NOT a Fragment. :)

^ Deleted the Fragment folder (package), and put it in .romstats

Enabled the Exit Menu in the ActionBar on the WebView in the case
users want to Exit the application after looking at the WebView.
I also added a long set of comments in the .romstats.AnonymousStats.java
to further explain what I exactly did.

Deleted the code that forced the WebView to load in Desktop mode,
as it isn't needed because the site loads the same in this situation
either way.

in xml/preview_data, I disabled all of the Preferences except for the one
that requires you to click it to load your DU version, as having the
other Preferences Enabled wasn't needed in this case

To further add to my putting the Exit in the ActionBar in ViewStats, I
had to change the logic of the Intent in AnonStats,java to force it to
Exit the entire application, instead of basically navigating up.

TO-DO:
Fix Whitespace
Implement the MenuDrawer(??)
Add a Dialog the very first time the user opens up this application,
because being nice is a good thing. When I have some free time ill do 2 and 3,
but the next patchset will feature stupid whitespace fixes, and it'll be good
to merge.

ps2:
Deleted the 'fragments' package. My bad om forgetting to in the first ps

Disabled HWA in PreviewStats  because lazy

Fixed whitespace. This is good to merge bro

Change-Id: Ib06915ff8d5e4058bad0adb4d858ed3333a804da